refactor: split tools/executor.ts into focused modules
